Add `isFeatureFlagActive` method
This adds a first pass of the `isFeatureFlagActive` method as a helper function.
When called, the method would check the current value of `isActive` for a specific feature.
