--- utils/api/featureFlags.ts
@@ -1,3 +1,0 @@
-export const isFeatureFlagActive = (featureName: string) => {
-  return true;
-};
