+++ utils/featureFlagHelpers.ts
@@ -1,0 +1,7 @@
+import { getFeatureFlags } from '../features';
+
+export const isFeatureFlagActive = (featureName: string) => {
+  const features = getFeatureFlags();
+
+  return features[featureName].isActive;
+};
